Move ngraph_catalog.h and .c outside the enable_variables folder
