+++ ngraph_bridge/enable_variable_ops/ngraph_enter_in_catalog.h
@@ -21,7 +21,7 @@
 
 #include "ngraph/ngraph.hpp"
 
-#include "ngraph_bridge/enable_variable_ops/ngraph_catalog.h"
+#include "ngraph_bridge/ngraph_catalog.h"
 
 using namespace std;
 namespace ng = ngraph;
